automation/adaptive_montecarlo.py
#!/usr/bin/env python3
"""
SRCL Elite – Adaptive Monte Carlo Engine
----------------------------------------
Performs stochastic simulations with adaptive variance tracking.
Sends progress updates to Gemini for real-time interpretation.
"""

import os, json, random, time
from datetime import datetime
from pathlib import Path
from google.cloud import aiplatform

# === CONFIG ===
RUNS = 5000
STEPS = 2000
OUTPUT_DIR = Path("results_montecarlo_v3")
GEMINI_ENABLED = True
PROJECT_ID = "canvas-sum-481614-f6"
LOCATION = "us-central1"

def simulate(alpha, beta, gamma, D):
    """Placeholder for your SRCL ripple simulation kernel."""
    coherence = max(0, 1 - abs(alpha + beta - gamma) * random.random())
    entropy = random.random() * (1 - coherence)
    return {"coherence": coherence, "entropy": entropy}

def adaptive_montecarlo():
    OUTPUT_DIR.mkdir(exist_ok=True)
    results = []
    start_time = time.time()
    
    for i in range(RUNS):
        alpha, beta, gamma, D = [random.random() for _ in range(4)]
        res = simulate(alpha, beta, gamma, D)
        res["alpha"], res["beta"], res["gamma"], res["D"] = alpha, beta, gamma, D
        res["run_id"], res["timestamp"] = i, datetime.utcnow().isoformat()
        results.append(res)
        
        if i % 100 == 0:
            path = OUTPUT_DIR / f"log_{i:05}.jsonl"
            with open(path, "w") as f:
                f.write("\n".join([json.dumps(r) for r in results[-100:]]))
            print(f"[{i}/{RUNS}] Logged to {path}")
            if GEMINI_ENABLED:
                interpret_with_gemini(results[-100:])
    
    elapsed = round(time.time() - start_time, 2)
    print(f"✅ Adaptive Monte Carlo complete ({RUNS} runs in {elapsed}s)")

def interpret_with_gemini(batch):
    """Ask Gemini for real-time pattern recognition."""
    try:
        aiplatform.init(project=PROJECT_ID, location=LOCATION)
        from vertexai.preview import generative_models as gm
        model = gm.GenerativeModel("gemini-pro")
        text = json.dumps(batch[-3:], indent=2)
        prompt = f"Analyze SRCL simulation patterns in this data:\n{text}"
        response = model.generate_content(prompt)
        print("🧠 Gemini Insight:", response.text[:300], "...")
    except Exception as e:
        print(f"[⚠️ Gemini error] {e}")

if __name__ == "__main__":
    adaptive_montecarlo()
